fix: allow underscores in formula name (#2205)
* fix: allow underscores in formula name

I was trying to create two formulas per release. One signifying the latest and
another with the version explicitly: formula.rb and formula@v0.0.0.rb

The filenames are fine but the formula name for the explicit version
was created as FormulaATv0.0.0 which isn't valid Ruby.

Looking at the function history I couldn't find why underscores were being replaced
since underscores are valid for Ruby identifiers. I opted to replace dots in a formula with
underscores

- https://github.com/goreleaser/goreleaser/commit/1399a396036ea246522a4cbae7aef376d6dbbda2
- https://ruby-doc.org/docs/ruby-doc-bundle/Manual/man-1.4/syntax.html#ident

* test: add test for version in formula name

I think underscores are removed to make the Formula name cleaner

* test: re-add old test